fix(viewer): round-10 tri-review findings
- notifEnable: dispatch {detail:false} on early-denied path so sibling
  AnnBellBtn instances sync when permission is already blocked (finding 2)
- viewer_watchlist: read prev.length inside functional updaters for
  addPlayer/addDojo to avoid stale-closure on maybeFirstAdd (finding 6)
- viewer_notifications.test: hoist installLS to module scope; consolidate
  beforeEach/afterEach into parent describe; extract makeLocalStorageMock /
  makeThrowingLocalStorageMock to new __tests__/test_helpers.js (findings 3-5)
- Update "already denied" test to expect the new dispatch call
